requirements-agent.txt
aiosmtplib==3.0.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.1.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.0
langchain-core==1.1.0
langchain-openai==1.1.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
xxhash==3.6.0
zstandard==0.25.0
